Use MUnit in RhoRoutesSuite
